@@ -1,16 +1,17 @@
 name: mc_dio
 description: Command of DIO request,Encapsulating each request as an object, the request supports both centralized and separated API usage
-version: 0.2.4
+version: 0.3.0
 # author: marcochen
 homepage: https://github.com/CZXBigBrother/MC_DIO
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  dio: ^3.0.10
+  dio: ^4.0.0
+  #^3.0.10
   # dio_cookie_manager: ^1.0.0
   # cookie_jar: ^1.0.1
   
